### 2017-03-19 #184 @arfon
:wave: @luk87. Could you take a look at the editor list and see which editor might be closest to your research domain? http://joss.theoj.org/about#editorial_board

Once we have an editor assigned we can start looking for a reviewer.
