### 2017-03-19 #184 @luk87
Dear @arfon I think the editor closest to my research domain could be Jake Vanderplas .
Thank you
