### 2017-03-19 #184 @arfon
> Dear @arfon I think the editor closest to my research domain could be Jake Vanderplas.

OK thanks. @jakevdp - any chance you could edit this submission for us?
