### 2017-03-22 #184 @jakevdp
Hi all – this is pretty far from my area of expertise, and I wouldn't even know where to begin in finding good reviewers for this (I haven't thought about Lie algebras since my second year of graduate school). I'd be happy to serve as editor, but @luk87 do you have any recommendations for appropriate people to review this? Thanks.
