### 2017-03-23 #184 @danielskatz
@luk87. As I've talked to a couple of people in this area and browsed the code itself, it seems to me that this is not really research software, but is more a GUI around a somewhat simple code, with the intent that this is mostly used for education.  But I may be completely wrong, as this is not my technical area at all.

Can you help me understand the research usage of this software?
