### 2017-03-23 #184 @luk87
Dear @danielskatz in several areas, as higher spin physics, supergravity, double filed theory or General relativity as an example it is common to deal with tensors carrying indices in one of the classical Lie algebras.  Although  it is trivial to compute the dimension of irreps for the algebra A_{n} it is not completely trivial to impose irreducible conditions on tensors when one is dealing with the other algebras.  In these cases with the code one could immediately get the dimension of the representation codified in the tensor since it is trivial to translate tensorial indices in Young tableaux.
