### 2017-03-23 #184 @danielskatz
thanks - you might want to include this (or some version of it) in the documents/text about the software
