### 2017-03-23 1d060cc
Update paper.md
