+++ README.md
@@ -2,7 +2,7 @@
 #YoungTab
 
 ## Summary
-YoungTab is a mathematical tool to compute the dimension of irreducible representations of the classical Lie algebras A, B, C, D, described by Young tableaux.  In particular for any Young tableau drawn in the drawing area and selected rank it will list the dimension of the irreducible representation associated to it for the four classical Lie algebras.  Due to the correspondence between Young tableaux and irreducible tensors, it is useful in finding the representation associated to tensors carrying indices in any of the Lie algebras mentioned above.  YoungTab has a wide range of application, from pure math to theoretical physics, where group theoretical structures and tensors are ubiquitous. 
+YoungTab is a mathematical tool to compute the dimension of irreducible representations of the classical Lie algebras A, B, C, D, described by Young tableaux. In particular for any Young tableau drawn in the drawing area and selected rank it will list the dimension of the irreducible representation associated to it for the four classical Lie algebras. Due to the correspondence between Young tableaux and irreducible tensors, it is useful in finding the representation associated to tensors carrying indices in any of the Lie algebras mentioned above. Although it is trivial to compute the dimension of irreps for the A-type algebra it is not completely trivial to impose irreducible conditions on tensors when one is dealing with the other classical Lie algebras. In these cases with YoungTab one could immediately compute the dimension of the representation codified in the tensor since it is trivial to translate tensorial indices in Young tableaux. YoungTab has a wide range of application, from pure math to theoretical physics, where group theoretical structures and tensors are ubiquitous.
 
 ## Installation
 YoungTab is a program written in C code thought for linux 64 and the GUI is based on the GTK toolkit. It could be directly downloaded and used by placing the bin file in a folder containing another folder with name glade containing in turn the logo and the glade file. The required libraries are gtk, gdk, cairo, gobject, glib and pthread.
