### 2017-03-31 #221 @danielskatz
Reviewer instructions

> Please work through the checklist at the start of this issue.
>
> If you need any further guidance/clarification take a look at the reviewer guidelines here http://joss.theoj.org/about#reviewer_guidelines
>
> Please make a publication recommendation at the end of your review
>
> Any questions, please ask for help by commenting on this issue! 🚀

In brief:

@jmichel7: your job now is to go through the checklist at top of the issue, and check off the items that are good.

If you have specific detailed issues with the submission/software, you can open an issue in the software repository (https://github.com/luk87/YoungTab, select Issues tab, click on the green "New issue" button) and put a comment here in this issue that refer to the URL of that issue.

If you have general issues about this as a paper, feel free to add comments here.

In either case, @luk87 (the submitting author) will be notified of the issues and then can respond to them.  But I will only be notified of comments here.

Once everything is done, add a comment saying so and making a recommendation for what to do next.

Thanks!!
