### 2017-04-28 #221 @jmichel7
I found the following issues within the reviewer questions:
- Version number: no version number is visible anywhere within the package.
- Example usage: there is none. However, the graphic interface is easily discoverable.
- Functionality documentation: there is none. It seems from the interface that computation is limited
 to groups of rank 10 or less. Such a limitation leaves the program useful to physicists, but makes it probably useless for mathematicians. Thus the target audience can only be physicists. Also, one would like to know with certainty that there can be no overflows in the computations (are large integers supported?)
- Automated tests: there are none
-Community guidelines: there are none either.

Altogether, the program is just a (rather pleasant) widget doing one very specific computation.
Its functionality is thus very limited. Writing it in C (the source does not contain a single comment) makes it very difficult for other people to modify/extend, but this choice of language was probably guided by the chosen graphical interface.

To summarize: a nice widget but without any future. A similar project with a future would choose another
implementation language and perhaps be part of a more general project for Lie groups and algebras, like the Lie package of Arjeh Cohen and al.
