### 2017-04-30 #221 @luk87
@jmichel7 and @danielskatz thank you. I try to clarify each point.

> I found the following issues within the reviewer questions:
>
>     Version number: no version number is visible anywhere within the package.
>     Example usage: there is none. However, the graphic interface is easily discoverable.

I can fix these issues immediately describing an example in the documentation. For the version number I put it in the release as required by GitHub. I will report it in the documentation if needed.

>     Functionality documentation: there is none. It seems from the interface that computation is limited
>     to groups of rank 10 or less. Such a limitation leaves the program useful to physicists, but makes it probably useless for mathematicians. Thus the target audience can only be physicists. Also, one would like to know with certainty that there can be no overflows in the computations (are large integers supported?)
>     Automated tests: there are none

I am going to integrate a control sequence to signal overflow. The overflow could appear in the computation of r and h in the code. I have defined these as floats. The limitation to rank 10 was imposed thinking to the usual problems it was supposed to solve. It eventually could be relaxed and also the computational power could be increased (clearly depending on how far one wants to go it requires different approaches). One way to do this is to look for simplification directly in the matrices r and h.

>     Community guidelines: there are none either.
>
> Altogether, the program is just a (rather pleasant) widget doing one very specific computation.
> Its functionality is thus very limited. Writing it in C (the source does not contain a single comment) makes it very difficult for other people to modify/extend, but this choice of language was probably guided by the chosen graphical interface.
>
> To summarize: a nice widget but without any future. A similar project with a future would choose another
> implementation language and perhaps be part of a more general project for Lie groups and algebras, like the Lie package of Arjeh Cohen and al.

The program is thought just as a simple tool, that can be useful also to people not working specifically on this subject and need to compute the dimension of tensors carrying indices of the classical Lie algebras, for example people working in gauge theory. Even if it was not built as part of a bigger project there could be some developments, as the inclusion of the computation of product of representations and the possibility to do branching of representations.
