+++ functions20170429b.h
@@ -1,0 +1,882 @@
+//We define the four functions computing the dimension of the tableaux, one for each classical Lie algebra  
+    
+    
+ void rank_function (GtkWidget *one,  gpointer user_data)
+{  
+  
+  data *ptr;
+  ptr= (data *) user_data;
+  char str_count[11] = {1};
+  ptr->num=gtk_combo_box_get_active (GTK_COMBO_BOX(ptr->combo1));
+  ptr->num+=1;
+  sprintf(str_count, "%d", ptr->num);
+  gtk_label_set_text(GTK_LABEL(ptr->label1), str_count ); 
+  gtk_label_set_text(GTK_LABEL(ptr->label2), str_count );
+  gtk_label_set_text(GTK_LABEL(ptr->label3), str_count );
+  gtk_label_set_text(GTK_LABEL(ptr->label4), str_count );
+ 
+ 
+  return ;  
+  
+}
+ 
+ 
+ 
+ //function computing the dimension of tableaux for algebra of type A
+ 
+ 
+ 
+  void func_A (GtkWidget *name, gpointer user_data)
+{
+   
+  
+  
+  tableaux *ptr;
+  ptr= (tableaux *) user_data;
+  
+  char str_count[10] = {0};
+  
+  
+  if(ptr->control == FALSE){
+    
+    gtk_label_set_text(GTK_LABEL(ptr->label1), "no" );
+    goto jump1;
+  };
+  
+  
+  
+  ptr->n=gtk_combo_box_get_active (GTK_COMBO_BOX(ptr->combo1));
+  
+  int n=ptr->n+1;
+  
+  if(n<= 0){
+    
+    gtk_label_set_text(GTK_LABEL(ptr->label1), "-" );
+    goto jump1;
+  };
+  
+int dim=0;    
+
+int **rvalue;
+int **hook;
+float h=1.0;
+float r=1.0;
+
+
+int ncolumn= ptr->cellinrow[0];
+int nrow=ptr->cellincolumn[0];
+
+
+int i=0;
+int j=0;
+int a=0;
+
+
+
+while(j < ncolumn){
+while(i < nrow){
+if(ptr->cellinrow[i]-j> 0){a+=1;};
+i+=1;  
+}
+
+ptr->cellincolumn[j]=a;
+a=0;
+i=0;
+j+=1;
+};
+
+j=0;
+i=0;
+
+
+  hook=(int **)malloc(nrow*sizeof(int *));/*righe*/  
+  for(i=0; i<nrow; ++i){ 
+    hook[i]=(int *) malloc(ptr->cellinrow[i]*sizeof(int));/*colonne*/
+   };
+
+
+  
+for(i=0; i<nrow; ++i){
+for(j=0; j<ptr->cellinrow[i]; ++j){
+hook[i][j]=ptr->cellinrow[i]+ptr->cellincolumn[j]-i-j-1;   
+//printf("%d\n",hook[i][j]);
+};
+};
+
+
+rvalue=(int **) malloc(nrow*sizeof(int *));/*righe*/  
+  for(i=0; i<nrow; ++i){ 
+    rvalue[i]=(int *) malloc(ncolumn*sizeof(int));/*colonne*/
+   };
+   
+for(i=0; i<nrow; ++i){
+for(j=0; j<ptr->cellinrow[i]; ++j){
+rvalue[i][j]=0;   
+};
+};
+   
+
+i=0;
+j=0;
+
+
+for(i=0; i<nrow; ++i){
[…]
