### 2017-04-30 #221 @danielskatz
@luk87 Please let us know when you have made these changes.

@arfon Can you comment on the issue of scope/weight:
> Altogether, the program is just a (rather pleasant) widget doing one very specific computation.
> Its functionality is thus very limited. Writing it in C (the source does not contain a single comment) makes it very difficult for other people to modify/extend, but this choice of language was probably guided by the chosen graphical interface.
>
> To summarize: a nice widget but without any future. A similar project with a future would choose another
> implementation language and perhaps be part of a more general project for Lie groups and algebras, like the Lie package of Arjeh Cohen and al.
