+++ README.md
@@ -11,3 +11,6 @@ YoungTab is a program written in C code thought for linux 64 and the GUI is base
 ## Usage
 The user could select the rank of the algebra from the list menu. On the display there are four different labels, under the column "algebra", for the four classical Lie algebras. The Young tableaux could be drawn on the blackboard, the black drawing area. Any left mouse click adds a cell justifying it automatically to the left. The dimension of the representations described by the Young tableaux appearing on the blackboard, is sketched near the corresponding algebra in the column "Dim". If the tableau does not respect the rules required to be identified with a legitimate Lie algebra irreducible representation (non increasing number of cells in each row from the top to the bottom and non increasing number of cells in each column from the left to the right) the output Dim will say "no". If the tableux exceeds the computational power it says "over". The tableaux could be deleted by a right mouse click on the blackboard.
 
+## Example
+As an example we can try to compute the independent components of the tensor T<sub>MN,P</sub> where MN are antisymmetric indices and are symmetrized with respect to P. This tensor is represented by a Young tableaux of type [2,1] namely with two cells in the first row and one cell in the second row. Depending on the type of algebra the indices belong the different irreducibility condition give a different number of indpendent components. Drawing the table it is immediate to get
+
