### 2017-04-30 #221 @arfon
> @arfon Can you comment on the issue of scope/weight:

@danielskatz - I think the (limited) scope is fine. Sounds like @jmichel7 is recommending it might be worth adding some documentation to the code to improve the legibility?
