### 2017-05-01 0dd3e7f
Update paper.md
