+++ README.md
@@ -15,3 +15,12 @@ The user could select the rank of the algebra from the list menu. On the display
 As an example we can try to compute the independent components of the tensor T<sub>MN,P</sub> where MN are antisymmetric indices and are symmetrized with respect to P and all the indices belong to a rank 4 algebra. This tensor is represented by a Young tableaux of type [2,1] namely with two cells in the first row and one cell in the second row. Depending on the type of algebra the indices belong, different irreducibility conditions should be imposed resulting in a different number of indpendent components. Drawing the table it is immediate to get the result for A<sub>4</sub> (SL(5)), B<sub>4</sub> (SO(9)), C<sub>4</sub> (Sp(8)) and D<sub>4</sub> (SO(8)).
 
 ![YoungTab example](https://github.com/luk87/YoungTab/blob/master/YTimage.png) 
+
+### Test
+
+Taking the Young tableuax [2,1] its dimension for the rank n classical Lie algebras should give the numerical value for the well known formulae
+
+* A<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    n(n+2)(n+1)/3
+* B<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    (2n+1)(n<sup>2</sup>-1)/3
+* C<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    8n(n<sup>2</sup>-1)/3
+* D<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    8n(n<sup>2</sup>-1)/3.
