+++ README.md
@@ -21,6 +21,6 @@ As an example we can try to compute the independent components of the tensor T<s
 Taking the Young tableuax [2,1] its dimension for the rank n classical Lie algebras should give the numerical value for the well known formulae
 
 * A<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    n(n+2)(n+1)/3
-* B<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    (2n+1)(n<sup>2</sup>-1)/3
+* B<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    (2n+3)(4n<sup>2</sup>-1)/3
 * C<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    8n(n<sup>2</sup>-1)/3
 * D<sub>n</sub> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;    8n(n<sup>2</sup>-1)/3.
