### 2017-05-01 45b834b
Update paper.md
