### 2017-05-01 3aef2c2
Update functions20170429b.h
