### 2017-05-01 a06f51f
Update functions20170429b.h
