### 2017-05-01 #221 @luk87
@danielskatz I have released the software with the overflow control and I have added to the documentation and example and a test to verify functionality. Furthermore I have added comments for a better reading of the source files, main20170429.c and functions20170429.h.
