### 2017-05-03 #221 @jmichel7
Most points I touched on have been improved. They are not yet perfect, but perhaps that's the end
of my comments.
