### 2017-05-03 #221 @jmichel7
I checked more boxes.
I did not find the test to verify functionality in the archive. Also there are still no community guidelines.
Is this last necessary?
