### 2017-05-03 #221 @luk87
@jmichel7 regarding the test I put a manual test in the documentation: the formula for the dimension of the tableaux [2,1] as function of the rank for the four algebras. Regarding the guidelines I assumed these were disciplined by the license and GitHub, but eventually I could add some more specific statements.
