### 2017-05-04 #221 @danielskatz
The idea of community guidelines is that the project should be welcoming of new contributors, and should explain how they can join in and contribute.  The text can be somewhat generic, but needs to be part of the project repo/documentation.
