### 2017-05-04 #221 @luk87
@danielskatz thank you for clarifying. I have added the guidelines in the README.
