### 2017-05-04 #221 @danielskatz
@jmichel7 please see if you are ok with the test and guidelines now
