### 2017-05-06 1b09047
Delete functions20170215.h
