+++ main20170215.c
@@ -1,416 +1,0 @@
-
-
-
-#include <gtk/gtk.h>
-#include <malloc.h>
-#include <glib.h>
-#include<stdio.h>
-#include<string.h>
-#include <math.h>
-#include <cairo.h>
-#include <unistd.h>
-#include <limits.h>
-#include "typedef.h"
-#include "functions20170215.h"
-
-
-
-
-
-int tabside=20;
-
-
-
-
-
-
-GtkWidget *g_rankA;
-GtkWidget *g_rankB;
-GtkWidget *g_rankC;
-GtkWidget *g_rankD;
-GtkWidget *g_dimA;
-GtkWidget *g_dimB;
-GtkWidget *g_dimC;
-GtkWidget *g_dimD;
-GtkWidget *g_combobox1;
-GtkTreeModel *g_list1;
-GtkWidget *g_button1;
-GtkWidget *g_area;
-GtkWidget *g_image2;
-
-static cairo_surface_t *surface = NULL;
-
-
-static void clear_surface (void)
-{
-  cairo_t *cr;
-  cr = cairo_create (surface);
-  cairo_set_source_rgb (cr, 0.3,0.3,0.3);
-  cairo_paint (cr);
-  cairo_destroy (cr);
-  
-  return;
-}
-
-
-static gboolean configure_event_cb (GtkWidget *widget, 
-                                    GdkEventConfigure *event, 
-                                    gpointer data)
-{
-  if (surface)
-    cairo_surface_destroy (surface);
-    surface = gdk_window_create_similar_surface (gtk_widget_get_window (widget),
-                                               CAIRO_CONTENT_COLOR,
-                                               gtk_widget_get_allocated_width (widget),
-                                               gtk_widget_get_allocated_height (widget));
-
-  
-  clear_surface ();
-
-  return TRUE;
-}
-
-
-
-static gboolean draw_cb (GtkWidget *widget,  
-                         cairo_t   *cr,  
-                         gpointer   data)
-{
-  cairo_set_source_surface (cr, surface, 0, 0);
-  cairo_paint (cr);
-
-  return FALSE;
-}
-
-
-static void draw_cell (GtkWidget *widget,  
-                        gdouble    x,  
-                        gdouble    y, 
-                        gpointer   user_data)
-{
-  
-    
-  tableaux *ptr;
-  ptr= (tableaux *) user_data;
-    
-  cairo_t *cr;
-
-  int nx;
-  int ny;
-  int k=0;
-  int l=0;
-  int i=0;
-  int j=0;
-  
-  
-  
-  int size=20;
-  
-  nx= (int) x/size;
-  ny= (int) y/size;
-  
-  int xmod= (int) (nx+1)*size;
-  int ymod= (int) (ny+1)*size;
-  
-  
-  
-  
-  
-  while(i<tabside){
-      if(ptr->ncell[i][ny+1]==88){
-      k+=1;};       
-      i+=1;};
-      //printf("%d %d\n",k,ny+1);     
-    ptr->ncell[k][ny+1]=88;
-    xmod=(k+2)*size;
-  
-    
- 
-    
- 
-ptr->control=TRUE;
-for(i=0;i<tabside;i++){
-k=0;
-for(j=0;j<tabside-1;j++){ 
-  
-if(ptr->ncell[i][j] == 88 && ptr->ncell[i][j+1] != 88) {k+=1;}
-  else;
-};
-if(k>1){ptr->control=FALSE;}
-else;
-};
-
-
-i=0;
-j=0;
-k=0;
-
-
-
-
[…]
