### 2017-05-08 #221 @arfon
@luk87 - as per the [author guidelines](http://joss.theoj.org/about#author_guidelines) your `paper.md` should only contain two sections (Summary/References). The other content you have here (Installation/Usage/Example) should be moved to either the repository README or some other docs.

In addition, please make sure you're citing the references in your bibtex file correctly as they won't be included otherwise. You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)

Please let me know when you've done this and I'll process the submission.
