### 2017-05-08 7c9ce86
Update paper.md
