### 2017-05-09 #221 @luk87
@arfon I have improved the paper and the references. However my affiliation in the meanwhile has changed. For the moment I have no affiliation, while when I submitted the paper I was affiliated with the Leibniz universität hannover. Could this be a problem and what affiliation should I specify in the paper?
