### 2017-05-09 #221 @arfon
> Could this be a problem and what affiliation should I specify in the paper?

I think it's OK to keep the affiliation the same. Have you completed your changes?
